--- a/internal/kgateway/extensions2/plugins/trafficpolicy/traffic_policy_plugin.go
+++ b/internal/kgateway/extensions2/plugins/trafficpolicy/traffic_policy_plugin.go
@@ -224,7 +224,12 @@ func registerTypes(ourCli versioned.Interface) {
 	)
 }
 
-func NewPlugin(ctx context.Context, commoncol *common.CommonCollections, mergeSettings string) extensionsplug.Plugin {
+func NewPlugin(
+	ctx context.Context,
+	commoncol *common.CommonCollections,
+	v validator.Validator,
+	mergeSettings string,
+) extensionsplug.Plugin {
 	registerTypes(commoncol.OurClient)
 
 	useRustformations = commoncol.Settings.UseRustFormations // stash the state of the env setup for rustformation usage
@@ -236,7 +241,6 @@ func NewPlugin(ctx context.Context, commoncol *common.CommonCollections, mergeSe
 	gk := wellknown.TrafficPolicyGVK.GroupKind()
 
 	translator := NewTrafficPolicyConstructor(ctx, commoncol)
-	v := validator.New()
 
 	// TrafficPolicy IR will have TypedConfig -> implement backendroute method to add prompt guard, etc.
 	policyCol := krt.NewCollection(col, func(krtctx krt.HandlerContext, policyCR *v1alpha1.TrafficPolicy) *ir.PolicyWrapper {