--- a/internal/kgateway/setup/setup.go
+++ b/internal/kgateway/setup/setup.go
@@ -33,6 +33,7 @@ import (
 	"github.com/kgateway-dev/kgateway/v2/pkg/pluginsdk/krtutil"
 	"github.com/kgateway-dev/kgateway/v2/pkg/settings"
 	"github.com/kgateway-dev/kgateway/v2/pkg/utils/envutils"
+	"github.com/kgateway-dev/kgateway/v2/pkg/validator"
 )
 
 type Server interface {
@@ -131,6 +132,12 @@ func WithGlobalSettings(settings *settings.Settings) func(*setup) {
 	}
 }
 
+func WithValidator(v validator.Validator) func(*setup) {
+	return func(s *setup) {
+		s.validator = v
+	}
+}
+
 type setup struct {
 	gatewayControllerName    string
 	gatewayClassName         string
@@ -148,6 +155,7 @@ type setup struct {
 	krtDebugger        *krt.DebugHandler
 	globalSettings     *settings.Settings
 	leaderElectionID   string
+	validator          validator.Validator
 }
 
 var _ Server = &setup{}
@@ -205,6 +213,10 @@ func New(opts ...func(*setup)) (*setup, error) {
 			slog.Error("error creating xds listener", "error", err)
 			return nil, err
 		}
+	}
+
+	if s.validator == nil {
+		s.validator = validator.NewBinaryValidator("")
 	}
 
 	return s, nil
@@ -283,7 +295,10 @@ func (s *setup) Start(ctx context.Context) error {
 
 	BuildKgatewayWithConfig(
 		ctx, mgr, s.gatewayControllerName, s.gatewayClassName, s.waypointClassName,
-		s.agentGatewayClassName, setupOpts, s.restConfig, istioClient, commoncol, agwCollections, uccBuilder, s.extraPlugins, s.extraAgentgatewayPlugins, s.extraGatewayParameters)
+		s.agentGatewayClassName, setupOpts, s.restConfig, istioClient, commoncol,
+		agwCollections, uccBuilder, s.extraPlugins, s.extraAgentgatewayPlugins,
+		s.extraGatewayParameters, s.validator,
+	)
 
 	slog.Info("starting admin server")
 	go admin.RunAdminServer(ctx, setupOpts)
@@ -313,6 +328,7 @@ func BuildKgatewayWithConfig(
 	extraPlugins func(ctx context.Context, commoncol *common.CommonCollections, mergeSettingsJSON string) []sdk.Plugin,
 	extraAgentgatewayPlugins func(ctx context.Context, agw *agentgatewayplugins.AgwCollections) []agentgatewayplugins.AgentgatewayPlugin,
 	extraGatewayParameters func(cli client.Client, inputs *deployer.Inputs) []deployer.ExtraGatewayParameters,
+	validator validator.Validator,
 ) error {
 	slog.Info("creating krt collections")
 	krtOpts := krtutil.NewKrtOptions(ctx.Done(), setupOpts.KrtDebugger)
@@ -344,6 +360,7 @@ func BuildKgatewayWithConfig(
 		KrtOptions:               krtOpts,
 		CommonCollections:        commonCollections,
 		AgwCollections:           agwCollections,
+		Validator:                validator,
 	})
 	if err != nil {
 		slog.Error("failed initializing controller: ", "error", err)