<!--
Thanks for opening a PR! Please delete any sections that don’t apply.
-->

# Description

<!--
A concise explanation of the change. You may include:
- **Motivation:** why this change is needed
- **What changed:** key implementation details
- **Related issues:** e.g., `Fixes #123`
-->

WIP: Optimize the gateway translator suite runtime. The TestRouteReplacement suite accounts for most of the performance because we need to enable STRICT validation mode, and because we're running this locally, we need to use the docker-based validator, which runs the envoy-gloo container image and shells out to envoy validate mode NxM where N is the number of sub-tests and M is the number of times translation calls the Validate method.

This commit explores two optimizations efforts:

1. Use a persistent docker that's managed in the RR suite
2. Wrap the validator instance and cache the results

For the former, this approach dramatically reduces system resource usage, but doesn't influence overall runtime in any meaningful way.

Before:

```bash
$ time go test -failfast ./internal/kgateway/translator/gateway
ok       github.com/kgateway-dev/kgateway/v2/internal/kgateway/translator/gateway        53.701s
go test -failfast ./internal/kgateway/translator/gateway  52.29s user 7.54s system 88% cpu 1:07.66 total
```

After:

```
$ time go test -failfast ./internal/kgateway/translator/gateway
ok       github.com/kgateway-dev/kgateway/v2/internal/kgateway/translator/gateway        55.206s
go test -failfast ./internal/kgateway/translator/gateway  12.36s user 2.24s system 25% cpu 57.973 total
```

That's ~12% runtime improvement. When wrapping the persistent docker-based validator with caching, we start to see considerable runtime improvements:

After:

```bash
$ time go test -failfast ./internal/kgateway/translator/gateway
ok       github.com/kgateway-dev/kgateway/v2/internal/kgateway/translator/gateway        27.217s
go test -failfast ./internal/kgateway/translator/gateway  23.69s user 3.34s system 80% cpu 33.583 total
```

# Change Type

<!--
Select one or more of the following by including the corresponding slash-command:
```
/kind breaking_change
/kind bug_fix
/kind design
/kind cleanup
/kind deprecation
/kind documentation
/kind flake
/kind new_feature
```
-->

/kind cleanup

# Changelog

<!--
Provide the exact line to appear in release notes for the chosen changelog type.

If no, just write "NONE" in the release-note block below.
If yes, a release note is required:
-->

```release-note
NONE
```

# Additional Notes

<!--
Any extra context or edge cases for reviewers.
-->